+++ packages/terra-date-picker/src/react-datepicker/time.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import classNames from 'classnames/bind'
 import {
   getHour,
   getMinute,
@@ -12,6 +13,9 @@ import {
   isTimeInDisabledRange,
   isTimeDisabled
 } from './date_utils'
+import styles from './stylesheets/datepicker.scss'
+
+const cx = classNames.bind(styles);
 
 export default class Time extends React.Component {
   static propTypes = {
@@ -77,13 +81,13 @@ export default class Time extends React.Component {
   }
 
   liClasses = (time, currH, currM) => {
-    let classes = ['react-datepicker__time-list-item']
+    let classes = ['react-datepicker-time-list-item']
 
     if ((currH === getHour(time)) && (currM === getMinute(time))) {
-      classes.push('react-datepicker__time-list-item--selected')
+      classes.push('react-datepicker-time-list-item--selected')
     }
     if (((this.props.minTime || this.props.maxTime) && isTimeInDisabledRange(time, this.props)) || (this.props.excludeTimes && isTimeDisabled(time, this.props.excludeTimes))) {
-      classes.push('react-datepicker__time-list-item--disabled')
+      classes.push('react-datepicker-time-list-item--disabled')
     }
 
     return classes.join(' ')
@@ -103,7 +107,7 @@ export default class Time extends React.Component {
     }
 
     return times.map((time, i) =>
-      <li key={i} onClick={this.handleClick.bind(this, time)} className={this.liClasses(time, currH, currM)}>
+      <li key={i} onClick={this.handleClick.bind(this, time)} className={cx(this.liClasses(time, currH, currM))}>
         {formatDate(time, format)}
       </li>
     )
@@ -116,13 +120,13 @@ export default class Time extends React.Component {
     }
 
     return (
-      <div className={`react-datepicker__time-container ${(this.props.todayButton) ? 'react-datepicker__time-container--with-today-button' : ''}`}>
-        <div className="react-datepicker__header react-datepicker__header--time">
-          <div className="react-datepicker-time__header">Time</div>
+      <div className={cx(['react-datepicker-time-container', `${(this.props.todayButton) ? 'react-datepicker-time-container--with-today-button' : ''}`])}>
+        <div className={cx(['react-datepicker-header', 'react-datepicker-header--time'])}>
+          <div className={cx('react-datepicker-time-header')}>Time</div>
         </div>
-        <div className="react-datepicker__time">
-          <div className="react-datepicker__time-box">
-            <ul className="react-datepicker__time-list" ref={list => { this.list = list }} style={height ? {height} : {}}>
+        <div className={cx('react-datepicker-time')}>
+          <div className={cx('react-datepicker-time-box')}>
+            <ul className={cx('react-datepicker-time-list')} ref={list => { this.list = list }} style={height ? {height} : {}}>
               {this.renderTimes.bind(this)()}
             </ul>
           </div>